backend/requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
pydantic==2.12.5
torch==2.9.1
transformers==5.3.0
safetensors==0.7.0
tokenizers==0.22.2
accelerate==1.13.0
huggingface-hub==1.4.1
yfinance==1.2.0
statsmodels==0.14.6
numpy==2.4.1
pandas==2.3.3
pyarrow==23.0.1
python-dotenv==1.2.1
python-multipart==0.0.21
